% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyAE.R
\name{tidyAE}
\alias{tidyAE}
\title{Read and Tidy Acoustic Emission Data}
\source{
https://stackoverflow.com/questions/79157266/parsing-and-working-with-microsecond-precision-timestamps-in-r-using-dplyr
}
\usage{
tidyAE(AEdata_filepath, xlsheet = "Line Display", skip_lines = 1)
}
\arguments{
\item{AEdata_filepath}{AEdata_filepath A string specifying the file path to the
Excel file containing AE data.}

\item{xlsheet}{xlsheet A string specifying the name of the Excel sheet to read.
Default is "Line Display".}

\item{skip_lines}{An integer specifying the number of lines to skip at the beginning of the Excel sheet.
Default is 1.}
}
\value{
A tibble containing the processed AE data
}
\description{
This function reads and processes Mistras Acoustic Emission (AE) data from an Excel file.
}
\details{
A tibble is returned with:

- Filtered to start from the row containing "ID"
- Filtered to remove rows where CH is not a number
- Date and time parsed from the original timestamp with microseconds processed
- Additional time-based columns added (DateDiff, WDay, Day, Hour)
- Filtered for a specific date range (2024-09-27 to 2024-10-15)
- Removed original timestamp column
- Reordered columns (character columns first, then numeric)
}
\examples{

\dontrun{
ae_data <- tidyAE("path/to/your/AEdata.xlsx")
}


}
